refactor(in-app): rename bottom/top sheet to bottom/top dialog
