+++ personalization-sdk/src/main/res/layout/bottom_dialog.xml
