+++ personalization-sdk/src/main/res/layout/top_dialog.xml
